Rename tests/goth to tests/goth_tests to avoid conflicts with goth pkg
