+++ pyproject.toml
@@ -105,9 +105,9 @@ ya-client-payment = "0.1.0"
 ya-market = "0.1.0"
 
 [tool.poe.tasks]
-test = "pytest --cov=yapapi --ignore tests/goth"
-goth-assets = "python -m goth create-assets tests/goth/assets"
-goth-tests = "pytest -svx tests/goth"
+test = "pytest --cov=yapapi --ignore tests/goth_tests"
+goth-assets = "python -m goth create-assets tests/goth_tests/assets"
+goth-tests = "pytest -svx tests/goth_tests"
 typecheck = "mypy ."
 codestyle = "black --check --diff ."
 _liccheck_export = "poetry export -E cli -f requirements.txt -o .requirements.txt"
