+++ tests/goth_tests/test_run_blender.py
@@ -10,11 +10,7 @@ from goth.runner.log import configure_logging
 from goth.runner import Runner
 from goth.runner.probe import RequestorProbe
 
-from tests.goth.assertions import (
-    assert_no_errors,
-    assert_all_invoices_accepted,
-    assert_tasks_processed,
-)
+from .assertions import assert_no_errors, assert_all_invoices_accepted, assert_tasks_processed
 
 
 logger = logging.getLogger("goth.test.run_blender")
