+++ tests/goth_tests/test_run_simple_service.py
@@ -11,7 +11,7 @@ from goth.runner.log import configure_logging
 from goth.runner import Runner
 from goth.runner.probe import RequestorProbe
 
-from tests.goth.assertions import assert_no_errors, assert_all_invoices_accepted
+from .assertions import assert_no_errors, assert_all_invoices_accepted
 
 
 logger = logging.getLogger("goth.test.run_simple_service")
